requirements.txt
agora_token_builder==1.0.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
cloudinary==1.41.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1.3
django-allauth==65.3.0
django-ckeditor-5==0.2.15
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-htmx==1.21.0
django-js-asset==2.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pillow==11.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
zope.interface==7.2
